updated validation
