+++ client/package.json
@@ -29,14 +29,9 @@
     "@tanstack/react-query": "^5.80.7",
     "@tiptap/extension-heading": "^2.23.0",
     "@tiptap/extension-history": "^2.23.0",
-    "@tiptap/extension-link": "^2.23.0",
     "@tiptap/extension-list-item": "^2.23.0",
     "@tiptap/extension-ordered-list": "^2.23.0",
     "@tiptap/extension-paragraph": "^2.23.0",
-    "@tiptap/extension-placeholder": "^2.23.1",
-    "@tiptap/extension-underline": "^2.23.0",
-    "@tiptap/react": "^2.23.0",
-    "@tiptap/starter-kit": "^2.23.0",
     "axios": "^1.10.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
