+++ client/src/app/routes/app/instructor/CourseFormPage.jsx
@@ -4,7 +4,7 @@ import CourseForm from "@/components/instructor/AddCourse/CourseForm";
 
 const CourseFormPage = () => {
   const { courseId } = useParams(); // Optional param
-
+console.log(courseId,"ci");
   return (
    
       <CourseForm courseId={courseId} />
